Fix dpkg-maintscript-helper calls
Reading again dpkg-maintscript-helper (1) man page I see I did not pay due
attention WRT prior-version: it needs to be based on the first version
removing the obsolete conffile, not on the last version containing it.
Amending it now means this is the first version properly removing them.

Closes: #669948
Thanks: Josh Triplett